refactor: consolidate the occupancy domain into one feature package #307
