+++ backend/src/main/java/com/occupi/config/InfluxLastCacheInitializer.java
@@ -1,7 +1,7 @@
 package com.occupi.config;
 
 import com.occupi.feature.database.repository.MetricsRepository;
-import com.occupi.feature.database.repository.OccupancyRepository;
+import com.occupi.feature.occupancy.OccupancyRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
