+++ backend/src/main/java/com/occupi/feature/occupancy/OccupancyService.java
@@ -1,7 +1,5 @@
-package com.occupi.feature.database.service;
+package com.occupi.feature.occupancy;
 
-import com.occupi.feature.database.model.OccupancyData;
-import com.occupi.feature.database.repository.OccupancyRepository;
 import jakarta.annotation.PreDestroy;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
