+++ backend/src/main/java/com/occupi/feature/occupancy/SensorDataService.java
@@ -1,0 +1,7 @@
+package com.occupi.feature.occupancy;
+
+import com.occupi.feature.occupancy.dto.SensorData;
+
+public interface SensorDataService {
+    void process(SensorData data);
+}
