+++ backend/src/main/java/com/occupi/feature/occupancy/dto/OccupancyResponse.java
@@ -1,4 +1,4 @@
-package com.occupi.feature.provider.dto;
+package com.occupi.feature.occupancy.dto;
 
 import java.time.Instant;
