--- backend/src/main/java/com/occupi/feature/receiver/SensorDataService.java
@@ -1,7 +1,0 @@
-package com.occupi.feature.receiver;
-
-import com.occupi.feature.receiver.dto.SensorData;
-
-public interface SensorDataService {
-    void process(SensorData data);
-}
