+++ backend/src/test/java/com/occupi/feature/chart/ChartHistoryInfluxIntegrationTest.java
@@ -4,8 +4,8 @@ import com.influxdb.v3.client.InfluxDBClient;
 import com.occupi.common.TimeSlots;
 import com.occupi.feature.chart.dto.HistoryPoint;
 import com.occupi.feature.chart.dto.HistoryResponse;
-import com.occupi.feature.database.model.OccupancyData;
-import com.occupi.feature.database.repository.OccupancyRepository;
+import com.occupi.feature.occupancy.OccupancyData;
+import com.occupi.feature.occupancy.OccupancyRepository;
 import com.occupi.feature.room.RoomService;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
