+++ backend/src/test/java/com/occupi/feature/occupancy/OccupancyServiceTest.java
@@ -1,7 +1,5 @@
-package com.occupi.feature.database.service;
+package com.occupi.feature.occupancy;
 
-import com.occupi.feature.database.model.OccupancyData;
-import com.occupi.feature.database.repository.OccupancyRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
